# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StudioFlow API",
        "description": "Software Agency & Project Delivery Platform — API-first backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local dev server"
        }
    ],
    "paths": {
...
        "/projects/{project}/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "List invoices for a project",
                "operationId": "f9669568567050bccbfbbf9f8514d70b",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Invoice"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create an invoice (Finance / Agency Manager)",
                "operationId": "40ae2d540d56359d4b20c476b4ab01e9",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "items"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "deposit",
                                            "milestone",
                                            "final",
                                            "subscription"
                                        ]
                                    },
                                    "milestone_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "tax": {
                                        "type": "number"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "unit_price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created as draft, totals computed server-side",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "View an invoice",
                "operationId": "e6f9c25a1265b8a65b66fb22b220750a",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail with items and payments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}/send": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Mark invoice as sent (draft→sent)",
                "operationId": "580251555ab8dc471ce4dfe951a39da6",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sent"
                    },
                    "422": {
                        "description": "Illegal transition from current status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}/cancel": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Cancel an invoice",
                "operationId": "abe1a3632f5e9379be6beafe1e4298d7",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}/pay": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Pay an invoice (owning Client only)",
                "operationId": "13fb86e37a199e50b3a8723e8dd3e7b2",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "method"
                                ],
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "card",
                                            "bank_transfer",
                                            "manual"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paid, Payment record created"
                    },
                    "422": {
                        "description": "Invoice not in sent/overdue status"
                    },
                    "403": {
                        "description": "Not the project's client"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}/refund": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Refund a paid invoice (Finance only)",
                "operationId": "e6f5f1ffb5a63e8e91cd67ce64ed24c5",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refunded"
                    },
                    "422": {
                        "description": "Invoice not in paid status"
                    },
                    "403": {
                        "description": "Client cannot refund"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Invoice": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "invoice_number": {
                        "type": "string",
                        "example": "INV-20260925-AB12CD"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "deposit",
                            "milestone",
                            "final",
                            "subscription"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "sent",
                            "paid",
                            "overdue",
                            "cancelled",
                            "refunded"
                        ]
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float"
                    },
                    "tax": {
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "description": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                },
                                "unit_price": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "line_total": {
                                    "type": "number",
                                    "format": "float"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token from /api/login, without the word 'Bearer'.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}